Make Item::Handle -> int explicit.

+++ Inventory.cpp
@@ -239,7 +239,7 @@ void Inventory::invent_sort ()
 		else
 		{
 			// Shrug!
-			return a < b;
+			return (int)a < (int)b;
 		}
 	};
